* Patrick's codec and mappings are now main codec * Gulli Johansen's original codec and mapping can be installed with install.sh option “johansen”
