--- src/mixer/maps/X-Touch Mini.remotemap
@@ -1,109 +1,0 @@
-Propellerhead Remote Mapping File
-File Format Version	1.0.0
-Control Surface Manufacturer	Behringer
-Control Surface Model	X-Touch Mini
-Map Version	1.5.1
-
-// ---------------------- Mixer -----------
-
-Scope	Propellerheads	Reason Master Section
-
-Define Group	Rotary Target	Level	Pan
-
-//	CS Item	Key	Remotable Item	Scale	Mode	Groups
-
-Map	Layer A Button		Rotary Target=Pan
-Map	Layer B Button		Rotary Target=Level
-
-Map	Master Fader		Master Level
-Map	Rotary 1		Channel 1 Level			Level
-Map	Rotary 2		Channel 2 Level			Level
-Map	Rotary 3		Channel 3 Level			Level
-Map	Rotary 4		Channel 4 Level			Level
-Map	Rotary 5		Channel 5 Level			Level
-Map	Rotary 6		Channel 6 Level			Level
-Map	Rotary 7		Channel 7 Level			Level
-Map	Rotary 8		Channel 8 Level			Level
-Map	Rotary 1		Channel 1 Pan			Pan
-Map	Rotary 2		Channel 2 Pan			Pan
-Map	Rotary 3		Channel 3 Pan			Pan
-Map	Rotary 4		Channel 4 Pan			Pan
-Map	Rotary 5		Channel 5 Pan			Pan
-Map	Rotary 6		Channel 6 Pan			Pan
-Map	Rotary 7		Channel 7 Pan			Pan
-Map	Rotary 8		Channel 8 Pan			Pan
-Map	Button 1		Channel 1 Solo
-Map	Button 2		Channel 2 Solo
-Map	Button 3		Channel 3 Solo
-Map	Button 4		Channel 4 Solo
-Map	Button 5		Channel 5 Solo
-Map	Button 6		Channel 6 Solo
-Map	Button 7		Channel 7 Solo
-Map	Button 8		Channel 8 Solo
-Map	Left Button		Channel 1 Mute
-Map	Right Button		Channel 2 Mute
-Map	Rewind Button		Channel 3 Mute
-Map	Fast Fwd Button		Channel 4 Mute
-Map	Loop Button		Channel 5 Mute
-Map	Stop Button		Channel 6 Mute
-Map	Play Button		Channel 7 Mute
-Map	Record Button		Channel 8 Mute
-
-// ---------------------- Combinator -----------
-
-Scope	Propellerheads	Combinator
-
-Define Group	Rotary Mode	Fine	Coarse
-
-//	CS Item	Key	Remotable Item	Scale	Mode	Groups
-
-Map	Layer A Button		Rotary Mode=Fine
-Map	Layer B Button		Rotary Mode=Coarse
-
-Map	Rotary 1		Enabled
-
-Map	Rotary 2		Pitch Bend	5		Coarse
-Map	Rotary 2		Pitch Bend	1		Fine
-
-Map	Rotary 3		Mod Wheel	5		Coarse
-Map	Rotary 3		Mod Wheel	1		Fine
-
-Map	Button 1		Run Pattern Devices
-Map	Left Button		Bypass All FX
-
-Map	Rotary 5		Rotary 1	5		Coarse
-Map	Rotary 5		Rotary 1	1		Fine
-Map	Button 5		Button 1
-
-Map	Rotary 6		Rotary 2	5		Coarse
-Map	Rotary 6		Rotary 2	1		Fine
-Map	Button 6		Button 2
-
-Map	Rotary 7		Rotary 3	5		Coarse
-Map	Rotary 7		Rotary 3	1		Fine
-Map	Button 7		Button 3
-
-Map	Rotary 8		Rotary 4	5		Coarse
-Map	Rotary 8		Rotary 4	1		Fine
-Map	Button 8		Button 4
-
-// ---------------------- Matrix -----------
-
-Scope	Propellerheads	Matrix Pattern Sequencer
-
-Map	Button 1		Pattern 1
-Map	Button 2		Pattern 2
-Map	Button 3		Pattern 3
-Map	Button 4		Pattern 4
-Map	Button 5		Pattern 5
-Map	Button 6		Pattern 6
-Map	Button 7		Pattern 7
-Map	Button 8		Pattern 8
-Map	Left Button		Bank A
-Map	Right Button		Bank B
-Map	Rewind Button		Bank C
-Map	Fast Fwd Button		Bank D
-Map	Stop Button		Pattern Enable
-Map	Play Button		Run
-Map	Rotary 8		Resolution
-
